style(canvas): add styles for context item and remove unused color picker component
- Imported SCSS styles for the ContextItem component to enhance its appearance.
- Added border-radius to the context preview popover for improved aesthetics.
- Removed the unused CommonColorPicker component to streamline the codebase.
