+++ packages/ai-workspace-common/src/components/canvas/launchpad/context-manager/index.scss
@@ -7,6 +7,7 @@
     background: transparent;
     box-shadow: none;
     padding: 0;
+    border-radius: 16px;
   }
 
   .ant-popover-inner-content {
