+++ packages/ai-workspace-common/src/components/canvas/nodes/shared/color-picker/index.scss
@@ -1,0 +1,7 @@
+.color-picker-popover {
+  .ant-popover-inner {
+    padding: 14px 8px;
+    border-radius: 12px;
+    background-color: var(--refly-bg-float-z3);
+  }
+}
